+++ src/combat/Player.cpp
@@ -157,7 +157,7 @@ void Player::heal(int amount) {
 
     bool wasBloodied = curHp <= maxHp/2;
 
-    curHp = std::min(maxHp, curHp + amount);
+    curHp = std::min(static_cast<int>(maxHp), curHp + amount);
 
     if (wasBloodied && curHp > maxHp/2 && hasRelic<RelicId::RED_SKULL>()) {
         decrementStatus<PS::STRENGTH>(3);
